feat: Event course API for owner

+++ src/db/course/leader.rs
@@ -5,7 +5,7 @@ use crate::common::Team;
 use crate::db::get_pool_conn;
 use crate::error::Error;
 
-pub fn sieve_list(course_id: u64) -> Result<Vec<(Team, bool)>, Error> {
+pub fn sieve_list(course_id: u32) -> Result<Vec<(Team, bool)>, Error> {
     let mut conn: PooledConn = get_pool_conn();
     let stmt = conn.prep(
         "SELECT t.team_id, t.team_key, t.name, t.description, cs.access
@@ -33,7 +33,7 @@ pub fn sieve_list(course_id: u64) -> Result<Vec<(Team, bool)>, Error> {
     Ok(teams)
 }
 
-pub fn sieve_edit(course_id: u64, team_id: u64, access: bool) -> Result<(), Error> {
+pub fn sieve_edit(course_id: u32, team_id: u64, access: bool) -> Result<(), Error> {
     let mut conn: PooledConn = get_pool_conn();
     let stmt = conn.prep(
         "INSERT INTO course_leader_sieves (course_id, team_id, access)
@@ -50,7 +50,7 @@ pub fn sieve_edit(course_id: u64, team_id: u64, access: bool) -> Result<(), Erro
     Ok(())
 }
 
-pub fn sieve_remove(course_id: u64, team_id: u64) -> Result<(), Error> {
+pub fn sieve_remove(course_id: u32, team_id: u64) -> Result<(), Error> {
     let mut conn: PooledConn = get_pool_conn();
     let stmt = conn.prep(
         "DELETE FROM course_leader_sieves
