+++ src/route/admin/course/mod.rs
@@ -36,7 +36,7 @@ pub fn course_create(session: UserSession, course: Json<Course>) -> Result<Strin
 }
 
 #[rocket::post("/admin/course_edit?<course_id>", format = "application/json", data = "<course>")]
-pub fn course_edit(session: UserSession, course_id: u64, course: Json<Course>) -> Result<(), Error> {
+pub fn course_edit(session: UserSession, course_id: u32, course: Json<Course>) -> Result<(), Error> {
     if !session.right.right_course_write {
         return Err(Error::RightCourseMissing);
     };
@@ -46,7 +46,7 @@ pub fn course_edit(session: UserSession, course_id: u64, course: Json<Course>) -
 }
 
 #[rocket::head("/admin/course_delete?<course_id>")]
-pub fn course_delete(session: UserSession, course_id: u64) -> Result<(), Error> {
+pub fn course_delete(session: UserSession, course_id: u32) -> Result<(), Error> {
     if !session.right.right_course_write {
         return Err(Error::RightCourseMissing);
     };
@@ -56,7 +56,7 @@ pub fn course_delete(session: UserSession, course_id: u64) -> Result<(), Error> 
 }
 
 #[rocket::get("/admin/course_event_list?<course_id>")]
-pub fn course_event_list(session: UserSession, course_id: u64) -> Result<Json<Vec<Event>>, Error> {
+pub fn course_event_list(session: UserSession, course_id: u32) -> Result<Json<Vec<Event>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
     };
@@ -75,7 +75,7 @@ pub fn course_event_list(session: UserSession, course_id: u64) -> Result<Json<Ve
 }
 
 #[rocket::get("/admin/course_requirement_list?<course_id>")]
-pub fn course_requirement_list(session: UserSession, course_id: u64) -> Result<Json<Vec<Requirement>>, Error> {
+pub fn course_requirement_list(session: UserSession, course_id: u32) -> Result<Json<Vec<Requirement>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
     };
@@ -85,7 +85,7 @@ pub fn course_requirement_list(session: UserSession, course_id: u64) -> Result<J
 }
 
 #[rocket::head("/admin/course_requirement_add?<course_id>&<skill_id>&<rank>")]
-pub fn course_requirement_add(session: UserSession, course_id: u64, skill_id: u32, rank: u32) -> Result<(), Error> {
+pub fn course_requirement_add(session: UserSession, course_id: u32, skill_id: u32, rank: u32) -> Result<(), Error> {
     if !session.right.right_course_write {
         return Err(Error::RightCourseMissing);
     };
@@ -107,7 +107,7 @@ pub fn course_requirement_remove(session: UserSession, requirement_id: u64) -> R
 #[rocket::get("/admin/course_statistic_class?<course_id>")]
 pub fn course_statistic_class(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
 ) -> Result<Json<Vec<(Event, u64, u64, u64)>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
@@ -120,7 +120,7 @@ pub fn course_statistic_class(
 #[rocket::get("/admin/course_statistic_leader?<course_id>")]
 pub fn course_statistic_leader(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
 ) -> Result<Json<Vec<(u64, String, String, u64)>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
@@ -133,7 +133,7 @@ pub fn course_statistic_leader(
 #[rocket::get("/admin/course_statistic_leader1?<course_id>&<leader_id>")]
 pub fn course_statistic_leader1(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
     leader_id: u64,
 ) -> Result<Json<Vec<(u64, String, NaiveDateTime, NaiveDateTime)>>, Error> {
     if !session.right.right_course_read {
@@ -147,7 +147,7 @@ pub fn course_statistic_leader1(
 #[rocket::get("/admin/course_statistic_supporter?<course_id>")]
 pub fn course_statistic_supporter(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
 ) -> Result<Json<Vec<(u64, String, String, u64)>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
@@ -160,7 +160,7 @@ pub fn course_statistic_supporter(
 #[rocket::get("/admin/course_statistic_supporter1?<course_id>&<supporter_id>")]
 pub fn course_statistic_supporter1(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
     supporter_id: u64,
 ) -> Result<Json<Vec<(u64, String, NaiveDateTime, NaiveDateTime)>>, Error> {
     if !session.right.right_course_read {
@@ -174,7 +174,7 @@ pub fn course_statistic_supporter1(
 #[rocket::get("/admin/course_statistic_participant?<course_id>")]
 pub fn course_statistic_participant(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
 ) -> Result<Json<Vec<(u64, String, String, u64)>>, Error> {
     if !session.right.right_course_read {
         return Err(Error::RightCourseMissing);
@@ -187,7 +187,7 @@ pub fn course_statistic_participant(
 #[rocket::get("/admin/course_statistic_participant1?<course_id>&<participant_id>")]
 pub fn course_statistic_participant1(
     session: UserSession,
-    course_id: u64,
+    course_id: u32,
     participant_id: u64,
 ) -> Result<Json<Vec<(u64, String, NaiveDateTime, NaiveDateTime)>>, Error> {
     if !session.right.right_course_read {
